+++ src/pages/_app.js
@@ -70,7 +70,6 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { WindowProvider } from 'src/windows'
 import { ErrorProvider } from 'src/error'
 import { useContext } from 'react'
-import TabBar from 'src/providers/TabBar'
 
 const clientSideEmotionCache = createEmotionCache()
 
@@ -113,9 +112,7 @@ const App = props => {
     (page => (
       <MenuProvider>
         <UserLayout contentHeightFixed={contentHeightFixed}>
-          <TabsProvider pageTitle={Component.pageTitle ?? pageProps.pageTitle}>
-            <TabBar>{page}</TabBar>
-          </TabsProvider>
+          <TabsProvider pageTitle={Component.pageTitle ?? pageProps.pageTitle}>{page}</TabsProvider>
         </UserLayout>
       </MenuProvider>
     ))
