+++ src/providers/TabsContext.js
@@ -1,12 +1,47 @@
 import React, { createContext, useEffect, useState, useContext } from 'react'
 import { useRouter } from 'next/router'
+import { Tabs, Tab, Box, IconButton, Menu, MenuItem } from '@mui/material'
+import CloseIcon from '@mui/icons-material/Close'
+import PropTypes from 'prop-types'
 import { MenuContext } from 'src/providers/MenuContext'
+import { RequestsContext } from './RequestsContext'
 
 const TabsContext = createContext()
 
+function CustomTabPanel(props) {
+  const { children, value, index, ...other } = props
+
+  return (
+    <Box
+      role='tabpanel'
+      id={`simple-tabpanel-${index}`}
+      aria-labelledby={`simple-tab-${index}`}
+      sx={{
+        display: value !== index ? 'none !important' : 'flex !important',
+        flexDirection: 'column',
+        width: '100%',
+        flex: '1 !important',
+        overflow: 'auto',
+        paddingTop: '5px',
+        position: 'relative',
+        backgroundColor: 'white'
+      }}
+      {...other}
+    >
+      {children}
+    </Box>
+  )
+}
+
+CustomTabPanel.propTypes = {
+  children: PropTypes.node,
+  index: PropTypes.number.isRequired,
+  value: PropTypes.number.isRequired
+}
+
 const TabsProvider = ({ children }) => {
   const router = useRouter()
-  const { menu, gear, lastOpenedPage } = useContext(MenuContext)
+  const { menu, gear, lastOpenedPage, reloadOpenedPage, setReloadOpenedPage } = useContext(MenuContext)
   const [anchorEl, setAnchorEl] = useState(null)
 
   const [openTabs, setOpenTabs] = useState([
@@ -19,6 +54,9 @@ const TabsProvider = ({ children }) => {
   const [currentTabIndex, setCurrentTabIndex] = useState(0)
   const [tabsIndex, setTabsIndex] = useState(null)
   const [initialLoadDone, setInitialLoadDone] = useState(false)
+  const { dashboardId } = JSON.parse(window.sessionStorage.getItem('userData'))
+
+  const open = Boolean(anchorEl)
 
   const OpenItems = (event, i) => {
     setTabsIndex(i)
@@ -91,6 +129,26 @@ const TabsProvider = ({ children }) => {
     }
   }
 
+  const reopenTab = tabRoute => {
+    const tabIndex = openTabs.findIndex(tab => tab.route === tabRoute)
+    if (tabIndex === -1) return
+    const tabData = openTabs[tabIndex]
+
+    const newOpenTabs = openTabs.filter((tab, index) => index !== tabIndex)
+    setOpenTabs(newOpenTabs)
+    if (tabIndex === currentTabIndex) {
+      setCurrentTabIndex(prev => (prev === 0 ? 0 : prev - 1))
+    } else if (tabIndex < currentTabIndex) {
+      setCurrentTabIndex(prev => prev - 1)
+    }
+    setTimeout(() => {
+      setOpenTabs(prev => [...prev.slice(0, tabIndex), tabData, ...prev.slice(tabIndex)])
+      setCurrentTabIndex(tabIndex)
+      router.push(tabRoute)
+    }, 0)
+    setReloadOpenedPage([])
+  }
+
   useEffect(() => {
     if (initialLoadDone) {
       const isTabOpen = openTabs.some((activeTab, index) => {
@@ -131,6 +189,7 @@ const TabsProvider = ({ children }) => {
   }, [children, router.asPath, initialLoadDone, lastOpenedPage, menu, gear])
 
   useEffect(() => {
+    if (router.asPath === reloadOpenedPage?.path + '/') reopenTab(reloadOpenedPage?.path + '/')
     if (!initialLoadDone && router.asPath && menu.length > 0) {
       const newTabs = [
         {
@@ -154,21 +213,138 @@ const TabsProvider = ({ children }) => {
       setOpenTabs(newTabs)
       setInitialLoadDone(true)
     }
-  }, [router.asPath, menu, gear, children, lastOpenedPage, initialLoadDone])
+  }, [router.asPath, menu, gear, children, lastOpenedPage, initialLoadDone, reloadOpenedPage])
 
-  const value = {
-    handleCloseOtherTab: () => {},
-    handleClose: () => {},
-    OpenItems: () => {},
-    handleChange: () => {},
-    currentTabIndex: 0,
-    anchorEl: null,
-    closeTab: () => {},
-    openTabs: [],
-    tabsIndex: null
-  }
-
-  return <TabsContext.Provider value={value}>{children}</TabsContext.Provider>
+  return (
+    <>
+      <Box
+        sx={{
+          display: 'flex !important',
+          flexDirection: 'column',
+          width: '100%',
+          flex: '1 !important',
+          overflow: 'auto'
+        }}
+      >
+        <Box sx={{ backgroundColor: '#231f20', pt: '5px', position: 'relative !important', zIndex: '3 !important' }}>
+          <Tabs
+            value={currentTabIndex}
+            onChange={handleChange}
+            variant='scrollable'
+            scrollButtons={openTabs.length > 3 ? 'auto' : 'off'}
+            aria-label='scrollable auto tabs example'
+            sx={{
+              minHeight: '35px !important',
+              '.MuiTab-root': {
+                color: 'white',
+                backgroundColor: 'grey',
+                '&:hover': {
+                  color: 'grey',
+                  backgroundColor: '#ddd'
+                }
+              },
+              '.Mui-selected': {
+                color: '#231f20',
+                backgroundColor: 'white'
+              },
+              '.MuiTabs-indicator': {
+                backgroundColor: 'white'
+              },
+              '.MuiSvgIcon-root': {
+                color: 'white!important'
+              },
+              '.MuiTab-root .MuiSvgIcon-root': {
+                color: '#5A585E !important'
+              }
+            }}
+          >
+            {openTabs.length > 0 &&
+              openTabs.map((activeTab, i) => (
+                <Tab
+                  key={i}
+                  label={activeTab?.label}
+                  onClick={() => {
+                    setCurrentTabIndex(i)
+                    router.push(activeTab.route)
+                  }}
+                  onContextMenu={event => OpenItems(event, i)}
+                  icon={
+                    activeTab.route === '/default/' ? null : (
+                      <IconButton
+                        size='small'
+                        onClick={event => {
+                          event.stopPropagation()
+                          closeTab(activeTab.route)
+                        }}
+                      >
+                        <CloseIcon fontSize='small' />
+                      </IconButton>
+                    )
+                  }
+                  iconPosition='end'
+                  sx={{
+                    minHeight: '35px !important',
+                    borderTopLeftRadius: 5,
+                    borderTopRightRadius: 5,
+                    py: '0px !important',
+                    mb: '0px !important',
+                    borderBottom: '0px !important',
+                    mr: '2px !important',
+                    fontWeight: '1.5rem',
+                    pr: '0px !important',
+                    pl: '10px !important',
+                    display: activeTab.route === '/default/' && dashboardId === null ? 'none' : 'flex'
+                  }}
+                />
+              ))}
+          </Tabs>
+        </Box>
+        {openTabs.length > 0 &&
+          openTabs.map((activeTab, i) => (
+            <CustomTabPanel key={activeTab.route} index={i} value={currentTabIndex}>
+              {activeTab.page}
+            </CustomTabPanel>
+          ))}
+      </Box>
+      <Menu
+        anchorEl={anchorEl}
+        id='account-menu'
+        open={open}
+        onClose={handleClose}
+        onClick={handleClose}
+        transformOrigin={{ horizontal: 'left', vertical: 'top' }}
+        anchorOrigin={{ horizontal: 'left', vertical: 'bottom' }}
+      >
+        <MenuItem
+          onClick={event => {
+            closeTab(openTabs[tabsIndex]?.route)
+            event.stopPropagation()
+            handleClose()
+          }}
+        >
+          <div>Close Tab</div>
+        </MenuItem>
+        <MenuItem
+          onClick={event => {
+            event.stopPropagation()
+            handleCloseOtherTab(tabsIndex)
+            handleClose()
+          }}
+        >
+          <div>Close Other Tabs</div>
+        </MenuItem>
+        <MenuItem
+          onClick={event => {
+            event.stopPropagation()
+            handleCloseAllTabs()
+            handleClose()
+          }}
+        >
+          <div>Close All Tabs</div>
+        </MenuItem>
+      </Menu>
+    </>
+  )
 }
 
 export { TabsContext, TabsProvider }
